requirements.txt
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.0
idna @ file:///croot/idna_1714398848350/work
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
munch==4.0.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
packaging==24.2
pandas==2.0.3
Pillow==9.3.0
pretrainedmodels==0.7.4
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.7.0
pytz==2024.2
requests @ file:///croot/requests_1721410876868/work
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.67.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
zipp==3.20.2
